[#12] Change .vue form adding css html js files and remove others files
