+++ puma/src/main.js
@@ -8,17 +8,11 @@ import store from './store';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import 'bootstrap-vue/dist/bootstrap-vue.min.css';
 
-import ModalDetalhamentoProjeto from './views/modal-detalhamento-projeto.vue';
-import BaseButton from './views/BaseButton.vue';
-
 import environment from './config/environment';
 
 Vue.config.productionTip = false;
 Vue.prototype.$http = axios;
 environment.configUser();
-
-Vue.component('modal-detalhamento-projeto', ModalDetalhamentoProjeto);
-Vue.component('base-button', BaseButton);
 
 Vue.use(BootstrapVue);
 Vue.use(IconsPlugin);
