+++ puma/src/views/cadastroProjeto/cadastro-projeto.css
@@ -1,0 +1,68 @@
+h2 {
+    margin-top: 50px;
+    margin-bottom: 20px;
+  }
+  h3{
+    font-size: 18px;
+    margin: 15px 0 3px 0;
+    position: relative;
+    display: inline-block;
+    font-weight: bold;
+  }
+  html{
+    margin-bottom: 27px;
+  }
+  input[type="checkbox"] {
+      margin: 0 4px 0 0;
+  }
+  div#checklist {
+    display: inline-flex;
+    padding-left: 13px;
+  }
+  label{
+    position: relative;
+  }
+  
+  input#file {
+      position: relative;
+      z-index: 5;
+      margin-bottom: 7px;
+  }
+  form {
+    margin-bottom: 25px;
+    margin-top: 32px;
+    width: 409px;
+    height: 200px;
+    border: 5px dashed darkblue;
+  }
+  form p {
+    width: 100%;
+    height: 100%;
+    text-align: center;
+    line-height: 39px;
+    color: red;
+    font-family: Helvetica, Arial, Verdana, Tahoma, sans-serif;
+    position: relative;
+    top: -173px;
+    font-size: 20px;
+  }
+  form input{
+    width: 100%;
+    height: 100%;
+    opacity: 0;
+  }
+  form label {
+      position: relative;
+      top: 7px;
+  }
+  #checklistArea {
+      margin: 10px 0 15px 0;
+      display: -webkit-inline-box;
+      position: relative;
+      padding-left: 8px;
+      left: -43px;
+  }
+  .btn{
+    color: black;
+    margin-left: 3%;
+  }
