+++ puma/src/views/cadastroProjeto/cadastro-projeto.html
@@ -1,0 +1,97 @@
+<div class="col-md-4 col-lg-offset-4">
+    <div class="form-group">
+      <h2 class="mb-10">{{operacao === 'visualizar'?'Visualização de Projeto':'Novo Projeto'}}</h2>
+      <div>
+        <input :style="{borderColor: !titulo.isValid ? 'red' : ''}" type="text"
+               v-model.trim="titulo.val" class="form-control"
+               id="titulo" placeholder="Qual é o título da proposta de projeto?"
+               @blur="validateFormInput('titulo')" @input="validateFormInput('titulo')"
+               :disabled = isLoading>
+        <p style="float: left" v-if="!titulo.isValid"
+           :style="{color: !titulo.val ? 'red' : ''}">Preenchimento obrigatório</p>
+      </div>
+    </div>
+
+     <div class="form-group">
+        <div class="">
+          <textarea
+              :style="{borderColor: !descricao.isValid ? 'red' : ''}"
+              v-model.trim="descricao.val" class="form-control" id="descricao"
+              placeholder="Qual a proposta do projeto?" rows="7" maxlength="10000"
+              @blur="validateFormInput('descricao')" @input="validateFormInput('descricao')"
+              :disabled = isLoading>
+          </textarea>
+          <p style="float: left" v-if="!descricao.isValid"
+             :style="{color: !descricao.val ? 'red' : ''}">
+            Preenchimento obrigatório
+          </p>
+        </div>
+     </div>
+
+    <div class="form-group" style="display: inline">
+      <div class="">
+        <textarea :style="{borderColor: !resultadoEsperado.isValid ? 'red' : ''}"
+                  v-model.trim="resultadoEsperado.val" class="form-control"
+                  id="descricao" placeholder="Qual resultado espera ter?" rows="5"
+                  @blur="validateFormInput('resultadoEsperado')"
+                  @input="validateFormInput('resultadoEsperado')"
+                  :disabled = isLoading>
+        </textarea>
+        <p style="position: relative; left: -117px; margin-bottom: 0;"
+           v-if="!resultadoEsperado.isValid"
+           :style="{color: !resultadoEsperado.val ? 'red' : ''}">
+          Preenchimento obrigatório
+        </p>
+      </div>
+    </div>
+
+    <div v-if="areasConhecimento.length">
+      <h3>
+        A quais áreas de conhecimento o projeto pertence?
+      </h3>
+    </div>
+    <div class="form-group"
+         v-for="areaConhecimento in areasConhecimento"
+         :key="areaConhecimento.id"
+         id="checklistArea"
+    >
+      <input :value="areaConhecimento.knowledgearea"
+             type="checkbox"
+             v-model="areasConhecimentoSelecionadas.val"
+             @change="validateFormInput('areaConhecimento')">
+      <label :for="areaConhecimento.id"> {{areaConhecimento.knowledgearea}}</label>
+    </div>
+
+    <div>
+      <label
+          id="negativeKnAreaForm"
+          style="margin-bottom: 10px;margin-bottom: 7px;position: relative;top: -11px;left: -33px;"
+          v-if="!areasConhecimentoSelecionadas.isValid"
+          :style="{color: !areasConhecimentoSelecionadas.isValid ? 'red' : ''}"
+          >
+        Selecione ao menos uma área de conhecimento
+      </label>
+    </div>
+
+    <div style="float: left" class="form-group" :style="{borderColor: !file.isValid ? 'red' : ''}">
+      <input
+          :disabled = isLoading
+          @change="updateFile($event)"
+          name="pic" id="file"
+          type="file"
+      >
+    </div>
+
+    <label for="file" class="mb-3"
+           :style="{color: !file.isValid ? 'red' : ''}"
+           v-if="!file.isValid"
+           style="position: relative;top: -16px;left: -100px;">
+      Preenchimento obrigatório
+    </label>
+      <div class="form-row" v-if="operacao !== 'visualizar'">
+        <button v-if="!isLoading"
+                type="button"
+                class="btn btn-warning"
+                @click="submitForm">Submeter</button>
+      </div>
+    </div>
