+++ puma/src/views/cadastroProjeto/cadastro-projeto.vue
@@ -1,0 +1,5 @@
+<template src="./cadastro-projeto.html"></template>
+
+<script src="./cadastro-projeto.js"></script>
+
+<style src="./cadastro-projeto.css"></style>
